modified:   0x02-functions_nested_loops/104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    unsigned int long long a, b, next;
+    unsigned int long a, b, next;
     int n;
 
     a = 1;
